Clean a lot of trash from example, Add touch wrapper
