+++ examples/current-sensor/.vscode/arduino.json
@@ -1,0 +1,5 @@
+{
+    "board": "esp8266:esp8266:d1",
+    "configuration": "CpuFrequency=80,FlashSize=4M1M,LwIPVariant=v2mss536,Debug=Disabled,DebugLevel=None____,FlashErase=none,UploadSpeed=921600",
+    "port": "COM18"
+}
