--- examples/current-sensor/I2CBitbanger.h
@@ -1,8 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-class I2CBitbanger {
-public:
-	I2CBitbanger(uint8_t sda, uint8_t sck);
-};
