--- examples/current-sensor/Version.h
@@ -1,1 +1,0 @@
-#pragma once
