+++ examples/current-sensor/src/common/CommonWifi.h
