+++ examples/current-sensor/src/common/Config.cpp
