+++ examples/current-sensor/src/common/Config.hpp
