+++ examples/current-sensor/src/common/Timeout.hpp
