+++ examples/current-sensor/src/common/UpdatableValue.hpp
